bloom_filter/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
redis[hiredis]==5.2.1
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
rich==13.9.4
mmh3==5.1.0
